+++ helpers_test.go
@@ -354,7 +354,6 @@ func Test_Utils_SortAcceptedTypes(t *testing.T) {
 		{spec: "text/html", quality: 1, specificity: 3, order: 0},
 		{spec: "text/*", quality: 0.5, specificity: 2, order: 1},
 		{spec: "*/*", quality: 0.1, specificity: 1, order: 2},
-		{spec: "application/json", quality: 0.999, specificity: 3, order: 3},
 		{spec: "application/xml", quality: 1, specificity: 3, order: 4},
 		{spec: "application/pdf", quality: 1, specificity: 3, order: 5},
 		{spec: "image/png", quality: 1, specificity: 3, order: 6},
@@ -363,8 +362,9 @@ func Test_Utils_SortAcceptedTypes(t *testing.T) {
 		{spec: "image/gif", quality: 1, specificity: 3, order: 9},
 		{spec: "text/plain", quality: 1, specificity: 3, order: 10},
 		{spec: "application/json", quality: 0.999, specificity: 3, params: headerParams{"a": []byte("1")}, order: 11},
+		{spec: "application/json", quality: 0.999, specificity: 3, order: 3},
 	}
-	sortAcceptedTypes(&acceptedTypes)
+	sortAcceptedTypes(acceptedTypes)
 	require.Equal(t, []acceptedType{
 		{spec: "text/html", quality: 1, specificity: 3, order: 0},
 		{spec: "application/xml", quality: 1, specificity: 3, order: 4},
@@ -390,7 +390,7 @@ func Benchmark_Utils_SortAcceptedTypes_Sorted(b *testing.B) {
 		acceptedTypes[0] = acceptedType{spec: "text/html", quality: 1, specificity: 1, order: 0}
 		acceptedTypes[1] = acceptedType{spec: "text/*", quality: 0.5, specificity: 1, order: 1}
 		acceptedTypes[2] = acceptedType{spec: "*/*", quality: 0.1, specificity: 1, order: 2}
-		sortAcceptedTypes(&acceptedTypes)
+		sortAcceptedTypes(acceptedTypes)
 	}
 	require.Equal(b, "text/html", acceptedTypes[0].spec)
 	require.Equal(b, "text/*", acceptedTypes[1].spec)
@@ -414,7 +414,7 @@ func Benchmark_Utils_SortAcceptedTypes_Unsorted(b *testing.B) {
 		acceptedTypes[8] = acceptedType{spec: "image/*", quality: 1, specificity: 2, order: 8}
 		acceptedTypes[9] = acceptedType{spec: "image/gif", quality: 1, specificity: 3, order: 9}
 		acceptedTypes[10] = acceptedType{spec: "text/plain", quality: 1, specificity: 3, order: 10}
-		sortAcceptedTypes(&acceptedTypes)
+		sortAcceptedTypes(acceptedTypes)
 	}
 	require.Equal(b, []acceptedType{
 		{spec: "text/html", quality: 1, specificity: 3, order: 0},
@@ -566,7 +566,7 @@ func Test_Utils_TestConn_Closed_Write(t *testing.T) {
 	require.NoError(t, err)
 
 	// Close early, write should fail
-	conn.Close() //nolint:errcheck, revive // It is fine to ignore the error here
+	conn.Close() //nolint:errcheck // It is fine to ignore the error here
 	_, err = conn.Write([]byte("Response 2\n"))
 	require.ErrorIs(t, err, errTestConnClosed)