+++ middleware/logger/logger_test.go
@@ -71,7 +71,7 @@ func Test_Logger(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${error}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	app.Get("/", func(_ fiber.Ctx) error {
@@ -94,7 +94,7 @@ func Test_Logger_locals(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${locals:demo}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	app.Get("/", func(c fiber.Ctx) error {
@@ -171,6 +171,147 @@ func Test_Logger_Done(t *testing.T) {
 	require.Positive(t, buf.Len(), 0)
 }
 
+// Test_Logger_Filter tests the Filter functionality of the logger middleware.
+// It verifies that logs are written or skipped based on the filter condition.
+func Test_Logger_Filter(t *testing.T) {
+	t.Parallel()
+
+	t.Run("Test Not Found", func(t *testing.T) {
+		t.Parallel()
+		app := fiber.New()
+
+		logOutput := bytes.Buffer{}
+
+		// Return true to skip logging for all requests != 404
+		app.Use(New(Config{
+			Skip: func(c fiber.Ctx) bool {
+				return c.Response().StatusCode() != fiber.StatusNotFound
+			},
+			Stream: &logOutput,
+		}))
+
+		resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/nonexistent", nil))
+		require.NoError(t, err)
+		require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
+
+		// Expect logs for the 404 request
+		require.Contains(t, logOutput.String(), "404")
+	})
+
+	t.Run("Test OK", func(t *testing.T) {
+		t.Parallel()
+		app := fiber.New()
+
+		logOutput := bytes.Buffer{}
+
+		// Return true to skip logging for all requests == 200
+		app.Use(New(Config{
+			Skip: func(c fiber.Ctx) bool {
+				return c.Response().StatusCode() == fiber.StatusOK
+			},
+			Stream: &logOutput,
+		}))
+
+		app.Get("/", func(c fiber.Ctx) error {
+			return c.SendStatus(fiber.StatusOK)
+		})
+
+		resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+		require.NoError(t, err)
+		require.Equal(t, fiber.StatusOK, resp.StatusCode)
+
+		// We skip logging for status == 200, so "200" should not appear
+		require.NotContains(t, logOutput.String(), "200")
+	})
+
+	t.Run("Always Skip", func(t *testing.T) {
+		t.Parallel()
+		app := fiber.New()
+
+		logOutput := bytes.Buffer{}
+
+		// Filter always returns true => skip all logs
+		app.Use(New(Config{
+			Skip: func(_ fiber.Ctx) bool {
+				return true // always skip
+			},
+			Stream: &logOutput,
+		}))
+
+		app.Get("/something", func(c fiber.Ctx) error {
+			return c.Status(fiber.StatusTeapot).SendString("I'm a teapot")
+		})
+
+		_, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/something", nil))
+		require.NoError(t, err)
+
+		// Expect NO logs
+		require.Empty(t, logOutput.String())
+	})
+
+	t.Run("Never Skip", func(t *testing.T) {
+		t.Parallel()
+		app := fiber.New()
+
+		logOutput := bytes.Buffer{}
+
+		// Filter always returns false => never skip logs
+		app.Use(New(Config{
+			Skip: func(_ fiber.Ctx) bool {
+				return false // never skip
+			},
+			Stream: &logOutput,
+		}))
+
+		app.Get("/always", func(c fiber.Ctx) error {
+			return c.Status(fiber.StatusTeapot).SendString("Teapot again")
+		})
+
+		_, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/always", nil))
+		require.NoError(t, err)
+
+		// Expect some logging - check any substring
+		require.Contains(t, logOutput.String(), strconv.Itoa(fiber.StatusTeapot))
+	})
+
+	t.Run("Skip /healthz", func(t *testing.T) {
+		t.Parallel()
+		app := fiber.New()
+
+		logOutput := bytes.Buffer{}
+
+		// Filter returns true (skip logs) if the request path is /healthz
+		app.Use(New(Config{
+			Skip: func(c fiber.Ctx) bool {
+				return c.Path() == "/healthz"
+			},
+			Stream: &logOutput,
+		}))
+
+		// Normal route
+		app.Get("/", func(c fiber.Ctx) error {
+			return c.SendString("Hello World!")
+		})
+		// Health route
+		app.Get("/healthz", func(c fiber.Ctx) error {
+			return c.SendString("OK")
+		})
+
+		// Request to "/" -> should be logged
+		_, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+		require.NoError(t, err)
+		require.Contains(t, logOutput.String(), "200")
+
+		// Reset output buffer
+		logOutput.Reset()
+
+		// Request to "/healthz" -> should be skipped
+		_, err = app.Test(httptest.NewRequest(fiber.MethodGet, "/healthz", nil))
+		require.NoError(t, err)
+		require.Empty(t, logOutput.String())
+	})
+}
+
 // go test -run Test_Logger_ErrorTimeZone
 func Test_Logger_ErrorTimeZone(t *testing.T) {
 	t.Parallel()
@@ -234,7 +375,7 @@ func Test_Logger_LoggerToWriter(t *testing.T) {
 
 			app.Use("/"+level, New(Config{
 				Format: "${error}",
-				Output: LoggerToWriter(logger, tc.
+				Stream: LoggerToWriter(logger, tc.
 					level),
 			}))
 
@@ -276,7 +417,7 @@ func Test_Logger_ErrorOutput_WithoutColor(t *testing.T) {
 	app := fiber.New()
 
 	app.Use(New(Config{
-		Output:        o,
+		Stream:        o,
 		DisableColors: true,
 	}))
 
@@ -293,7 +434,7 @@ func Test_Logger_ErrorOutput(t *testing.T) {
 	app := fiber.New()
 
 	app.Use(New(Config{
-		Output: o,
+		Stream: o,
 	}))
 
 	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
@@ -312,7 +453,7 @@ func Test_Logger_All(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${pid}${reqHeaders}${referer}${scheme}${protocol}${ip}${ips}${host}${url}${ua}${body}${route}${black}${red}${green}${yellow}${blue}${magenta}${cyan}${white}${reset}${error}${reqHeader:test}${query:test}${form:test}${cookie:test}${non}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	// Alias colors
@@ -324,6 +465,124 @@ func Test_Logger_All(t *testing.T) {
 
 	expected := fmt.Sprintf("%dHost=example.comhttpHTTP/1.10.0.0.0example.com/?foo=bar/%s%s%s%s%s%s%s%s%sCannot GET /", os.Getpid(), colors.Black, colors.Red, colors.Green, colors.Yellow, colors.Blue, colors.Magenta, colors.Cyan, colors.White, colors.Reset)
 	require.Equal(t, expected, buf.String())
+}
+
+func Test_Logger_CLF_Format(t *testing.T) {
+	t.Parallel()
+	buf := bytebufferpool.Get()
+	defer bytebufferpool.Put(buf)
+
+	app := fiber.New()
+
+	app.Use(New(Config{
+		Format: CommonFormat,
+		Stream: buf,
+	}))
+
+	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/?foo=bar", nil))
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
+
+	expected := fmt.Sprintf("0.0.0.0 - - [%s] \"%s %s %s\" %d %d\n",
+		time.Now().Format("15:04:05"),
+		fiber.MethodGet, "/?foo=bar", "HTTP/1.1",
+		fiber.StatusNotFound,
+		0)
+	logResponse := buf.String()
+	require.Equal(t, expected, logResponse)
+}
+
+func Test_Logger_Combined_CLF_Format(t *testing.T) {
+	t.Parallel()
+	buf := bytebufferpool.Get()
+	defer bytebufferpool.Put(buf)
+
+	app := fiber.New()
+
+	app.Use(New(Config{
+		Format: CombinedFormat,
+		Stream: buf,
+	}))
+	const expectedUA = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/74.0.3729.169 Safari/537.36"
+	const expectedReferer = "http://example.com"
+	req := httptest.NewRequest(fiber.MethodGet, "/?foo=bar", nil)
+	req.Header.Set("Referer", expectedReferer)
+	req.Header.Set("User-Agent", expectedUA)
+	resp, err := app.Test(req)
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
+
+	expected := fmt.Sprintf("0.0.0.0 - - [%s] %q %d %d %q %q\n",
+		time.Now().Format("15:04:05"),
+		fmt.Sprintf("%s %s %s", fiber.MethodGet, "/?foo=bar", "HTTP/1.1"),
+		fiber.StatusNotFound,
+		0,
+		expectedReferer,
+		expectedUA)
+	logResponse := buf.String()
+	require.Equal(t, expected, logResponse)
+}
+
+func Test_Logger_Json_Format(t *testing.T) {
+	t.Parallel()
+	buf := bytebufferpool.Get()
+	defer bytebufferpool.Put(buf)
+
+	app := fiber.New()
+
+	app.Use(New(Config{
+		Format: JSONFormat,
+		Stream: buf,
+	}))
+
+	req := httptest.NewRequest(fiber.MethodGet, "/?foo=bar", nil)
+	resp, err := app.Test(req)
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
+
+	expected := fmt.Sprintf(
+		"{\"time\":%q,\"ip\":%q,\"method\":%q,\"url\":%q,\"status\":%d,\"bytesSent\":%d}\n",
+		time.Now().Format("15:04:05"),
+		"0.0.0.0",
+		fiber.MethodGet,
+		"/?foo=bar",
+		fiber.StatusNotFound,
+		0,
+	)
+	logResponse := buf.String()
+	require.Equal(t, expected, logResponse)
+}
+
+func Test_Logger_ECS_Format(t *testing.T) {
+	t.Parallel()
+	buf := bytebufferpool.Get()
+	defer bytebufferpool.Put(buf)
+
+	app := fiber.New()
+
+	app.Use(New(Config{
+		Format: ECSFormat,
+		Stream: buf,
+	}))
+
+	req := httptest.NewRequest(fiber.MethodGet, "/?foo=bar", nil)
+	resp, err := app.Test(req)
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
+
+	expected := fmt.Sprintf(
+		"{\"@timestamp\":%q,\"ecs\":{\"version\":\"1.6.0\"},\"client\":{\"ip\":%q},\"http\":{\"request\":{\"method\":%q,\"url\":%q,\"protocol\":%q},\"response\":{\"status_code\":%d,\"body\":{\"bytes\":%d}}},\"log\":{\"level\":\"INFO\",\"logger\":\"fiber\"},\"message\":%q}\n",
+		time.Now().Format("15:04:05"),
+		"0.0.0.0",
+		fiber.MethodGet,
+		"/?foo=bar",
+		"HTTP/1.1",
+		fiber.StatusNotFound,
+		0,
+		fmt.Sprintf("%s %s responded with %d", fiber.MethodGet, "/?foo=bar", fiber.StatusNotFound),
+	)
+	logResponse := buf.String()
+	require.Equal(t, expected, logResponse)
 }
 
 func getLatencyTimeUnits() []struct {
@@ -358,7 +617,7 @@ func Test_Logger_WithLatency(t *testing.T) {
 	app := fiber.New()
 
 	logger := New(Config{
-		Output: buff,
+		Stream: buff,
 		Format: "${latency}",
 	})
 	app.Use(logger)
@@ -403,7 +662,7 @@ func Test_Logger_WithLatency_DefaultFormat(t *testing.T) {
 	app := fiber.New()
 
 	logger := New(Config{
-		Output: buff,
+		Stream: buff,
 	})
 	app.Use(logger)
 
@@ -453,7 +712,7 @@ func Test_Query_Params(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${queryParams}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/?foo=bar&baz=moz", nil))
@@ -474,7 +733,7 @@ func Test_Response_Body(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${resBody}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	app.Get("/", func(c fiber.Ctx) error {
@@ -508,7 +767,7 @@ func Test_Request_Body(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${bytesReceived} ${bytesSent} ${status}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	app.Post("/", func(c fiber.Ctx) error {
@@ -536,7 +795,7 @@ func Test_Logger_AppendUint(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${bytesReceived} ${bytesSent} ${status}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	app.Get("/", func(c fiber.Ctx) error {
@@ -611,7 +870,7 @@ func Test_Response_Header(t *testing.T) {
 	}))
 	app.Use(New(Config{
 		Format: "${respHeader:X-Request-ID}",
-		Output: buf,
+		Stream: buf,
 	}))
 	app.Get("/", func(c fiber.Ctx) error {
 		return c.SendString("Hello fiber!")
@@ -634,7 +893,7 @@ func Test_Req_Header(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${reqHeader:test}",
-		Output: buf,
+		Stream: buf,
 	}))
 	app.Get("/", func(c fiber.Ctx) error {
 		return c.SendString("Hello fiber!")
@@ -658,7 +917,7 @@ func Test_ReqHeader_Header(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${reqHeader:test}",
-		Output: buf,
+		Stream: buf,
 	}))
 	app.Get("/", func(c fiber.Ctx) error {
 		return c.SendString("Hello fiber!")
@@ -689,7 +948,7 @@ func Test_CustomTags(t *testing.T) {
 				return output.WriteString(customTag)
 			},
 		},
-		Output: buf,
+		Stream: buf,
 	}))
 	app.Get("/", func(c fiber.Ctx) error {
 		return c.SendString("Hello fiber!")
@@ -713,7 +972,7 @@ func Test_Logger_ByteSent_Streaming(t *testing.T) {
 
 	app.Use(New(Config{
 		Format: "${bytesReceived} ${bytesSent} ${status}",
-		Output: buf,
+		Stream: buf,
 	}))
 
 	app.Get("/", func(c fiber.Ctx) error {
@@ -724,7 +983,7 @@ func Test_Logger_ByteSent_Streaming(t *testing.T) {
 			for {
 				i++
 				msg := fmt.Sprintf("%d - the time is %v", i, time.Now())
-				fmt.Fprintf(w, "data: Message: %s\n\n", msg) //nolint:errcheck // ignore error
+				fmt.Fprintf(w, "data: Message: %s\n\n", msg)
 				err := w.Flush()
 				if err != nil {
 					break
@@ -759,7 +1018,7 @@ func Test_Logger_EnableColors(t *testing.T) {
 	app := fiber.New()
 
 	app.Use(New(Config{
-		Output: o,
+		Stream: o,
 	}))
 
 	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
@@ -782,7 +1041,7 @@ func Benchmark_Logger(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${bytesReceived} ${bytesSent} ${status}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Set("test", "test")
@@ -794,7 +1053,7 @@ func Benchmark_Logger(b *testing.B) {
 	b.Run("DefaultFormat", func(bb *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Hello, World!")
@@ -805,7 +1064,7 @@ func Benchmark_Logger(b *testing.B) {
 	b.Run("DefaultFormatDisableColors", func(bb *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
-			Output:        io.Discard,
+			Stream:        io.Discard,
 			DisableColors: true,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
@@ -819,7 +1078,7 @@ func Benchmark_Logger(b *testing.B) {
 		logger := fiberlog.DefaultLogger()
 		logger.SetOutput(io.Discard)
 		app.Use(New(Config{
-			Output: LoggerToWriter(logger, fiberlog.LevelDebug),
+			Stream: LoggerToWriter(logger, fiberlog.LevelDebug),
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Hello, World!")
@@ -831,7 +1090,7 @@ func Benchmark_Logger(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${bytesReceived} ${bytesSent} ${status} ${reqHeader:test}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Set("test", "test")
@@ -844,7 +1103,7 @@ func Benchmark_Logger(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${locals:demo}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Locals("demo", "johndoe")
@@ -857,7 +1116,7 @@ func Benchmark_Logger(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${locals:demo}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/int", func(c fiber.Ctx) error {
 			c.Locals("demo", 55)
@@ -874,7 +1133,7 @@ func Benchmark_Logger(b *testing.B) {
 					io.Discard.Write(logString) //nolint:errcheck // ignore error
 				}
 			},
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/logging", func(ctx fiber.Ctx) error {
 			return ctx.SendStatus(fiber.StatusOK)
@@ -886,7 +1145,7 @@ func Benchmark_Logger(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${pid}${reqHeaders}${referer}${scheme}${protocol}${ip}${ips}${host}${url}${ua}${body}${route}${black}${red}${green}${yellow}${blue}${magenta}${cyan}${white}${reset}${error}${reqHeader:test}${query:test}${form:test}${cookie:test}${non}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Hello, World!")
@@ -898,7 +1157,7 @@ func Benchmark_Logger(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${bytesReceived} ${bytesSent} ${status}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Set("Connection", "keep-alive")
@@ -908,7 +1167,7 @@ func Benchmark_Logger(b *testing.B) {
 				for {
 					i++
 					msg := fmt.Sprintf("%d - the time is %v", i, time.Now())
-					fmt.Fprintf(w, "data: Message: %s\n\n", msg) //nolint:errcheck // ignore error
+					fmt.Fprintf(w, "data: Message: %s\n\n", msg)
 					err := w.Flush()
 					if err != nil {
 						break
@@ -927,7 +1186,7 @@ func Benchmark_Logger(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${resBody}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Sample response body")
@@ -950,7 +1209,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${bytesReceived} ${bytesSent} ${status}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Set("test", "test")
@@ -962,7 +1221,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 	b.Run("DefaultFormat", func(bb *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Hello, World!")
@@ -975,7 +1234,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		logger := fiberlog.DefaultLogger()
 		logger.SetOutput(io.Discard)
 		app.Use(New(Config{
-			Output: LoggerToWriter(logger, fiberlog.LevelDebug),
+			Stream: LoggerToWriter(logger, fiberlog.LevelDebug),
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Hello, World!")
@@ -986,7 +1245,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 	b.Run("DefaultFormatDisableColors", func(bb *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
-			Output:        io.Discard,
+			Stream:        io.Discard,
 			DisableColors: true,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
@@ -999,7 +1258,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${bytesReceived} ${bytesSent} ${status} ${reqHeader:test}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Set("test", "test")
@@ -1012,7 +1271,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${locals:demo}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Locals("demo", "johndoe")
@@ -1025,7 +1284,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${locals:demo}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/int", func(c fiber.Ctx) error {
 			c.Locals("demo", 55)
@@ -1042,7 +1301,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 					io.Discard.Write(logString) //nolint:errcheck // ignore error
 				}
 			},
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/logging", func(ctx fiber.Ctx) error {
 			return ctx.SendStatus(fiber.StatusOK)
@@ -1054,7 +1313,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${pid}${reqHeaders}${referer}${scheme}${protocol}${ip}${ips}${host}${url}${ua}${body}${route}${black}${red}${green}${yellow}${blue}${magenta}${cyan}${white}${reset}${error}${reqHeader:test}${query:test}${form:test}${cookie:test}${non}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Hello, World!")
@@ -1066,7 +1325,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${bytesReceived} ${bytesSent} ${status}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			c.Set("Connection", "keep-alive")
@@ -1076,7 +1335,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 				for {
 					i++
 					msg := fmt.Sprintf("%d - the time is %v", i, time.Now())
-					fmt.Fprintf(w, "data: Message: %s\n\n", msg) //nolint:errcheck // ignore error
+					fmt.Fprintf(w, "data: Message: %s\n\n", msg)
 					err := w.Flush()
 					if err != nil {
 						break
@@ -1095,7 +1354,7 @@ func Benchmark_Logger_Parallel(b *testing.B) {
 		app := fiber.New()
 		app.Use(New(Config{
 			Format: "${resBody}",
-			Output: io.Discard,
+			Stream: io.Discard,
 		}))
 		app.Get("/", func(c fiber.Ctx) error {
 			return c.SendString("Sample response body")